requirements.txt
python-dotenv>=0.21.0
python-binance==1.0.15
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
streamlit>=1.25
plotly>=5.0
requests
torch>=2.0
joblib
pandas_ta
